Collect HirDisplay impls to a single file
